requirements.txt
Django>=5.0,<5.2
channels>=4.1
daphne>=4.1
Pillow>=10.0
python-dotenv>=1.0
djangorestframework>=3.15
django-widget-tweaks>=1.5
